feat(host-react-native): package skeleton + extract pure utilities (Phase 2a)
First slice of @tlsn/host-react-native. Scaffolds the new package and
moves the React-Native-flavored utility files that have no React Native
dependencies of their own (just import names like ViewStyle and TextStyle
from react-native types) — the safest pieces to extract first.

Package
- packages/host-react-native/ with package.json, tsconfig (jsx: react-jsx,
  module: ESNext, DOM in lib for the styles types), .gitignore.
- Exposes three submodule entry points:
  - @tlsn/host-react-native/logger
  - @tlsn/host-react-native/style
  - @tlsn/host-react-native/util
- Peer deps on react, react-native, react-native-webview;
  @react-native-cookies/cookies is peer-optional.

Files moved (mobile → host-react-native)
- lib/cssToRN.ts + cssToRN.test.ts → src/style/
- lib/logStore.ts + lib/installLogCapture.ts → src/logger/
- lib/oauthDetection.ts + oauthDetection.test.ts → src/util/

Mobile consumers (LogList.tsx, _layout.tsx, NativeProver.tsx,
PluginRenderer.tsx) now import from @tlsn/host-react-native/{logger,style}
instead of relative paths. app/mobile/package.json gains the workspace
dep; vitest config gets passWithNoTests since the only two test files
moved out (rebuilds when the bigger components extract land).

Verification: host-contracts builds, host-cli 5/5 tests pass,
host-react-native 33/33 tests pass (the migrated suites), mobile passes
(no tests).

Subsequent slices: MobilePluginHost, NativeProver, PluginRenderer,
PluginWebView, then refactor app/mobile to consume the full barrel.
