+++ app/mobile/package.json
@@ -24,6 +24,7 @@
     "@react-native-cookies/cookies": "^6.2.1",
     "@react-navigation/native": "^7.1.8",
     "@tlsn/host-contracts": "*",
+    "@tlsn/host-react-native": "*",
     "@tlsn/plugins": "*",
     "@types/uuid": "^10.0.0",
     "expo": "~54.0.33",
