+++ app/mobile/vitest.config.ts
@@ -7,6 +7,10 @@ export default defineConfig({
     environment: 'node',
     include: ['**/*.test.ts'],
     exclude: ['node_modules', 'android', 'ios', '.expo'],
+    // Mobile no longer ships its own test suite — those tests moved to
+    // @tlsn/host-react-native with the extracted utilities. Allow the
+    // run to pass when nothing matches.
+    passWithNoTests: true,
   },
   resolve: {
     alias: {
