+++ packages/host-react-native/.gitignore
@@ -1,0 +1,1 @@
+/dist/
