+++ packages/host-react-native/src/logger/installLogCapture.ts
