+++ packages/host-react-native/src/util/oauthDetection.test.ts
