+++ packages/host-react-native/src/util/oauthDetection.ts
